Failing test showing AST plugins "bleed" through all instances.
This test demonstrates how each "layer" in an ember app is transpiling
its own templates. Each is _supposed_ to be able to provide its own AST
plugins, **but** this test shows how this falls down and fails.

The current mechanism for registering AST plugins is fundamentally
flawed. This forced us (long ago) to add some shenanigans to clear the
require cache (to avoid this exact state leakage).