Avoid conflicts with ember-cli-htmlbars-inline-precompile
Both ember-cli-htmlbars-inline-precompile and ember-cli-htmlbars used
the same _exact_ check to determine if
`babel-plugin-htmlbars-inline-precompile` was registered. When the build
was parallelizable we would properly detect that the
babel-plugin-htmlbars-inline-precompile was or was not the one added by
each respective addon (because the addon's own `__dirname` is included
in the check) **but** when the build was not parallelizable the babel
plugins array contains a standard Babel plugins array. In that case,
both addons were _only_ checking if the plugin _path_ matched (via
`require.resolve`) which was completely insufficient.

This issue resulted in `ember-cli-htmlbars-inline-precompile` _thinking_
that it was already registered _when_ the build was not parallelizable,
which would subsequently make it not register its own instance of the
plugin (to support import paths other than `ember-cli-htmlbars`).

The fix here s twofold:

1. This forces any instances of ember-cli-htmlbars-inline-precompile
   addon to be completely inert if it is part of the same parent addon
   or app as ember-cli-htmlbars itself.  In addition, we emit a
   deprecation notice when applicable.
2. Updating the `_isInlinePrecompileBabelPluginRegistered` method to
   properly check if the plugin ember-cli-htmlbars is registering is
   already registered (with our configuration).