Enable colocated component classes to be TypeScript
Prior to this, we only looked for `*.js` files as the "backing class"
and would actually create a `.js` file right next to the existing `.ts`
file. This leads to a whole host of bugs for TypeScript users. :(

This updates the logic to look for `.js` file first then fall back to
`.ts` files.