Only check semver range if ember-source is present
Prior to this patch the value `false` would be passed to `semver.gte(`
causing an exception of `Invalid Version: false`. This ensures the range
validity of the version is only checked if one is present.

Specifically bower versions of Ember have an `emberVersion` of `false`
here. However none of those versions of Ember support co-location.

Fixes #422