Ensure custom templateCompilerPath is an absolute path.
The examples that we show for using a custom template compiler path use
a relative path, but that doesn't actually work when inline compilation
runs and is parallelized (because the worker processes run from a
different relative root inside `/tmp`).

This ensures that any specified overrides are converted to absolute
paths.