Ensure AST plugins have the same ordering as < ember-cli-htmlbars@5.5.0.
We have to reverse these for reasons that are a bit bonkers. The initial
version of this system used `registeredPlugin` from
`ember-template-compiler.js` to set up these plugins (because Ember ~
1.13 only had `registerPlugin`, and there was no way to pass plugins
directly to the call to `compile`/`precompile`). Calling
`registerPlugin` unfortunately **inverted** the order of plugins (it
essentially did `PLUGINS = [plugin, ...PLUGINS]`).

Sooooooo...... we are forced to maintain that **absolutely bonkers**
ordering.