Avoid repeated encoding in getTemplateCompiler
`crypto.createHash('md5').update(...)` takes a buffer. The prior code would read the file as a string _then_ convert back to a buffer. Now we only convert to a string for `new vm.Script()` which _requires_ a string. tldr; we convert back and forth from a buffer to a string one less time with the code change here.