Export non-user-constructible `TemplateFactory` type
This allows users who need to name the type (e.g. for custom wiring
with `setComponentTemplate`) to do so, while making it so that the type
can *only* be constructed by calling `hbs()` in type terms *and* in the
documentation for the item. Additionally, drop type-level compatibility
with the export from `htmlbars-inline-precompile`, which was deprecated
years ago, long before official TS support in this library.